+++ src/pyastgrep/asts.py
@@ -2,22 +2,61 @@
 from __future__ import annotations
 
 import ast
+import re
+import sys
 from functools import partial
 from typing import Callable
 
 from lxml import etree
 from lxml.etree import _Element
 
+illegal_unichrs = [
+    (0x00, 0x08),
+    (0x0B, 0x0C),
+    (0x0E, 0x1F),
+    (0x7F, 0x84),
+    (0x86, 0x9F),
+    (0xFDD0, 0xFDDF),
+    (0xFFFE, 0xFFFF),
+]
+if sys.maxunicode >= 0x10000:  # not narrow build
+    illegal_unichrs.extend(
+        [
+            (0x1FFFE, 0x1FFFF),
+            (0x2FFFE, 0x2FFFF),
+            (0x3FFFE, 0x3FFFF),
+            (0x4FFFE, 0x4FFFF),
+            (0x5FFFE, 0x5FFFF),
+            (0x6FFFE, 0x6FFFF),
+            (0x7FFFE, 0x7FFFF),
+            (0x8FFFE, 0x8FFFF),
+            (0x9FFFE, 0x9FFFF),
+            (0xAFFFE, 0xAFFFF),
+            (0xBFFFE, 0xBFFFF),
+            (0xCFFFE, 0xCFFFF),
+            (0xDFFFE, 0xDFFFF),
+            (0xEFFFE, 0xEFFFF),
+            (0xFFFFE, 0xFFFFF),
+            (0x10FFFE, 0x10FFFF),
+        ]
+    )
 
-def _set_encoded_literal(set_fn: Callable[[str | bytes], None], literal: bool | int | str | None) -> None:
+illegal_ranges = [rf"{chr(low)}-{chr(high)}" for (low, high) in illegal_unichrs]
+xml_illegal_character_regex = "[" + "".join(illegal_ranges) + "]"
+illegal_xml_chars_re = re.compile(xml_illegal_character_regex)
+
+
+def _set_encoded_literal(set_fn: Callable[[str | bytes], None], literal: bool | int | bytes | str | None) -> None:
     if isinstance(literal, (bool, int, float)):
         literal = str(literal)
     if literal is None:
         literal = ""
-    try:
-        set_fn(literal)
-    except ValueError:
-        set_fn("")  # Null byte - failover to empty string
+    if isinstance(literal, bytes):
+        literal = literal.decode("utf-8", errors="replace")
+
+    # NUL characters and control characters are not allowed in XML. It's better
+    # to be able to search the rest of the string, so we replace
+    set_fn(illegal_xml_chars_re.sub("", literal))
 
 
 def ast_to_xml(