+++ tests/examples/test_xml/bytes.py
@@ -1,0 +1,1 @@
+MYBYTES = b"hello"